Tidy up comments from PR
